move more funcs to helper util
